+++ Python/tdw/add_ons/replicant.py
@@ -105,7 +105,8 @@ class Replicant(ReplicantBase, ReplicantStatic, ReplicantDynamic):
                              animation=animation,
                              library=library,
                              collision_avoidance_distance=self._record.collision_avoidance_distance,
-                             collision_avoidance_half_extents=self._record.collision_avoidance_half_extents)
+                             collision_avoidance_half_extents=self._record.collision_avoidance_half_extents,
+                             collision_avoidance_y=self._record.collision_avoidance_y)
 
     def move_to(self, target: TARGET, reset_arms: bool = True, reset_arms_duration: float = 0.25,
                 scale_reset_arms_duration: bool = True, arrived_at: float = 0.1, bounds_position: str = "center",
@@ -147,7 +148,8 @@ class Replicant(ReplicantBase, ReplicantStatic, ReplicantDynamic):
                              animation=animation,
                              library=library,
                              collision_avoidance_distance=self._record.collision_avoidance_distance,
-                             collision_avoidance_half_extents=self._record.collision_avoidance_half_extents)
+                             collision_avoidance_half_extents=self._record.collision_avoidance_half_extents,
+                             collision_avoidance_y=self._record.collision_avoidance_y)
 
     def reach_for(self, target: Union[TARGET, List[TARGET]], arm: Union[Arm, List[Arm]], absolute: bool = True,
                   offhand_follows: bool = False, arrived_at: float = 0.09, max_distance: float = 1.5,